refactor: make filenames more useful, removed testing code, moved ArmState to utils
