+++ src/main/java/org/team1540/robot2023/commands/arm/ExtensionCommand.java
@@ -2,11 +2,11 @@ package org.team1540.robot2023.commands.arm;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class ExtensionPID extends CommandBase {
+public class ExtensionCommand extends CommandBase {
     private final Arm arm;
     private final double targetExtension;
 
-    public ExtensionPID(Arm arm, double targetExtension) {
+    public ExtensionCommand(Arm arm, double targetExtension) {
         this.arm = arm;
         this.targetExtension = targetExtension;
         addRequirements(arm);
