+++ src/main/java/org/team1540/robot2023/commands/grabber/GrabberIntakeCommand.java
@@ -3,10 +3,10 @@ package org.team1540.robot2023.commands.grabber;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import org.team1540.robot2023.Constants;
 
-public class GrabberIntake extends CommandBase {
+public class GrabberIntakeCommand extends CommandBase {
     private final WheeledGrabber wheeledGrabber;
 
-    public GrabberIntake(WheeledGrabber wheeledGrabber) {
+    public GrabberIntakeCommand(WheeledGrabber wheeledGrabber) {
         this.wheeledGrabber = wheeledGrabber;
         addRequirements(wheeledGrabber);
     }
