+++ src/main/java/org/team1540/robot2023/commands/grabber/GrabberOuttakeCommand.java
@@ -2,10 +2,10 @@ package org.team1540.robot2023.commands.grabber;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class GrabberOuttake extends CommandBase {
+public class GrabberOuttakeCommand extends CommandBase {
     private final WheeledGrabber wheeledGrabber;
 
-    public GrabberOuttake(WheeledGrabber wheeledGrabber){
+    public GrabberOuttakeCommand(WheeledGrabber wheeledGrabber){
         this.wheeledGrabber = wheeledGrabber;
         addRequirements(wheeledGrabber);
     }
